requirements.txt
torch
torchvision
pillow
matplotlib
numpy
tqdm
flask
